Versión 23-09

+++ ciclo3backGrupo13/src/main/java/co/edu/unbosque/ciclo3backGrupo13/dao/UsuariosDAO.java
@@ -1,9 +1,8 @@
 package co.edu.unbosque.ciclo3backGrupo13.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import co.edu.unbosque.ciclo3backGrupo13.model.Usuarios;
 
-public interface UsuariosDAO extends JpaRepository <Usuarios, Integer> {
+public interface UsuariosDAO extends JpaRepository <Usuarios, Long> {
 
 }
